Accept hostname targets that do not contain a period.
This fixes issue #123, which was due to a misinterpretation of the
line

`if host.split(".")[0][0].isalpha() or host.split(".")[-1][-1].isalpha():`

(see commit 8d46d1473111bfe3aba1e9a6e8a6bd39abbacf06, L156). Because of
how Python's str.split() works, a hostname that does not contain a
period when .split(".") results in a single-item list, so this
expression evaluates to True. However, this also means that the
.split()s had no effect: the line was equivalent to

`if host[0].isalpha() or host[-1].isalpha()`.